Added error handling, more responsiveness, max-width, checkoxes align in chrome, import fixed, and refactored
